added kick, leave and raffle endpionts to groups

+++ groups/views.py
@@ -3,9 +3,11 @@ from flask_login import current_user, login_required
 from loguru import logger
 
 from secret_santa import db
+from secret_santa.models import Group, User, Member, Santa
 from secret_santa.groups.forms import GroupCreateForm, GroupDeleteForm
-from secret_santa.models import Group, Member, User
+from secret_santa.models import Group, Member, User, Santa
 from secret_santa.utils import assign_all_santas, assign_santa_to_target
+
 
 # start application definitions
 groups = Blueprint("groups", __name__, template_folder="templates")
@@ -40,8 +42,6 @@ def create_post():
     group_form = GroupCreateForm(request.form)
     if group_form.validate():
         db.session.add(Group(name=request.form.get("name"), owner_id=current_user.id))
-        db.session.commit()
-
         group = Group.query.filter_by(name=request.form.get("name")).first()
         db.session.add(Member(group_id=group.id, user_id=current_user.id, is_owner=True))
         db.session.commit()
@@ -57,15 +57,18 @@ def create_post():
 @login_required
 def delete(group_id):
     found_group = Group.query.get(group_id)
-    if found_group:
+    if found_group and current_user.id == found_group.owner_id:
         db.session.delete(found_group)
+        found_members = Member.query.filter_by(group_id=group_id).all()
+        for member in found_members:
+            db.session.delete(member)
         db.session.commit()
 
         flash(f"Group <{group_id}> successfully deleted!", "is-success")
         return redirect(url_for("groups.index"))
 
-    flash(f"Couldn't find group <{group_id}>!", "is-warning")
-    return redirect(url_for("groups.index"))
+    flash(f"Could not delete Group <{group_id}>!", "is-warning")
+    return redirect(url_for('groups.profile', group_id=group_id))
 
 
 @groups.route("/<int:group_id>/profile")
@@ -75,11 +78,28 @@ def profile(group_id):
         flash("Could not find Group!", "is-warning")
         redirect(url_for("groups.index"))
 
-    users = []
-    for member in Member.query.filter_by(group_id=group_id).all():
-        users.append((User.query.get(member.user_id), member.is_owner))
+    users = (
+        Member.query.filter_by(group_id=group_id)
+        .join(User, Member.user_id == User.id)
+        .add_columns(User.id, User.name)
+        .all()
+    )
 
-    return render_template("groups/profile.html", group=found_group, users=users)
+    presentees = (
+        Member.query
+        .join(Santa, Member.user_id == Santa.santa_id, isouter=True)
+        .filter_by(group_id=group_id)
+        .join(User, Santa.santa_id == User.id, isouter=True)
+        .add_columns(User.id, User.name)
+        .all()
+    )
+
+    if not presentees:
+        presentees = [None] * len(users)
+
+    logger.debug(f"users: {users}")
+    logger.debug(f"presentees: {presentees}")
+    return render_template("groups/profile.html", group=found_group, users=zip(users, presentees))
 
 
 @groups.route("/<int:group_id>/join")
@@ -104,19 +124,53 @@ def join(group_id):
 @groups.route("/<int:group_id>/leave")
 @login_required
 def leave(group_id):
-    return "Sorry, the page was not implemneted yet!"
+    found_group = Group.query.get(group_id)
+    logger.warning(f"group: {found_group}")
+    member = Member.query.filter_by(group_id=group_id, user_id=current_user.id).first()
+    if found_group and member:
+        db.session.delete(member)
+        db.session.commit()
+
+        flash(f"Left Group <{group_id}> successfully!", "is-success")
+        return redirect(url_for('groups.profile', group_id=group_id))
+
+    flash(f"Could not remove User <{current_user.id}> from Group <{group_id}>!", "is-warning")
+    return redirect(url_for('groups.profile', group_id=group_id))
 
 
 @groups.route("/<int:group_id>/kick/<int:user_id>")
 @login_required
 def kick(group_id, user_id):
-    return "Sorry, the page was not implemneted yet!"
+    found_group = Group.query.get(group_id)
+    logger.warning(f"group: {found_group}")
+    member = Member.query.filter_by(group_id=group_id, user_id=user_id).first()
+    if found_group and member:
+        db.session.delete(member)
+        db.session.commit()
+
+        flash(f"User <{user_id}> was kicked from Group <{group_id}>!", "is-success")
+        return redirect(url_for('groups.profile', group_id=group_id))
+
+    flash(f"Could not remove User <{user_id}> from Group <{group_id}>!", "is-warning")
+    return redirect(url_for('groups.profile', group_id=group_id))
 
 
-@groups.route("/<int:group_id>/draw_init")
+@groups.route("/<int:group_id>/raffle")
 @login_required
 def raffle(group_id):
-    return "Sorry, the page was not implemneted yet!"
+    found_santas = Santa.query.filter_by(group_id=group_id).first()
+    if found_santas:
+        flash("Already raffled members!", "is-success")
+        return redirect(url_for('groups.profile', group_id=group_id))
+
+    found_group = Group.query.filter_by(id=group_id, owner_id=current_user.id).first()
+    if found_group:
+        assign_all_santas(found_group.id)
+        flash("Succesfully raffled members!", "is-success")
+        return redirect(url_for('groups.profile', group_id=group_id))
+
+    flash("Could not initiate raffle!", "is-warning")
+    return redirect(url_for('groups.profile', group_id=group_id))
 
 
 @groups.route("/<int:group_id>/draw")
